Fix infinite loading
