+++ packages/ui/src/hooks/dapp/use-entity-details.ts
@@ -4,9 +4,24 @@ import { useQueries } from '@tanstack/react-query'
 
 import { splitArrayIntoChunks } from 'ui/src/utils/array-chunk'
 
-import { defaultFungiblesOptIns, defaultNonFungiblesOptIns } from './use-entity-balances'
 import { useGatewayClient } from './use-gateway-client'
 import { useNetworkId } from './use-network'
+
+export const defaultFungiblesOptIns: StateEntityDetailsOptIns = {
+	ancestor_identities: false,
+	component_royalty_vault_balance: false,
+	package_royalty_vault_balance: false,
+	non_fungible_include_nfids: true,
+	explicit_metadata: ['name', 'symbol', 'description', 'tags', 'icon_url', 'info_url', 'validator', 'pool'],
+}
+
+export const defaultNonFungiblesOptIns: StateEntityDetailsOptIns = {
+	ancestor_identities: false,
+	component_royalty_vault_balance: false,
+	package_royalty_vault_balance: false,
+	non_fungible_include_nfids: true,
+	explicit_metadata: ['name', 'description', 'key_image_url'],
+}
 
 const defaultOptIns: StateEntityDetailsOptIns = {
 	ancestor_identities: true,
