+++ packages/ui/src/hooks/dapp/use-entity-nft.ts
@@ -12,14 +12,19 @@ export const useNonFungibleCollection = (resourceId: string) => {
 
 	return useInfiniteQuery({
 		queryKey: ['useNonFungibleCollection', networkId, resourceId],
-		queryFn: async ({ pageParam: cursor }) =>
+		queryFn: async ({ pageParam: lastPage }) =>
 			state.innerClient.nonFungibleIds({
 				stateNonFungibleIdsRequest: {
 					resource_address: resourceId,
-					cursor,
+					cursor: lastPage?.non_fungible_ids.next_cursor || null,
+					at_ledger_state: lastPage?.non_fungible_ids.next_cursor
+						? {
+								state_version: lastPage.ledger_state.state_version,
+						  }
+						: null,
 				},
 			}),
-		getNextPageParam: lastPage => lastPage.non_fungible_ids.next_cursor,
+		getNextPageParam: lastPage => (lastPage.non_fungible_ids.next_cursor ? lastPage : undefined),
 		enabled: !!resourceId && !!state,
 	})
 }
@@ -34,14 +39,15 @@ export const useNonFungibleIds = (resourceId: string, addresses: string[]) => {
 		queryFn: async ({ pageParam }) => {
 			const responses = await Promise.all(
 				vaults.map(({ account, vault, resource }, idx) =>
-					pageParam?.[idx] === null
-						? { items: [], next_cursor: null }
+					pageParam?.[idx].next_cursor === null
+						? { items: [], next_cursor: null, ledger_state: null }
 						: state.innerClient.entityNonFungibleIdsPage({
 								stateEntityNonFungibleIdsPageRequest: {
 									address: account,
 									resource_address: resource,
 									vault_address: vault,
-									cursor: pageParam?.[idx],
+									cursor: pageParam?.[idx].next_cursor || null,
+									at_ledger_state: pageParam?.[idx].next_cursor ? pageParam?.[idx].ledger_state : null,
 								},
 						  }),
 				),
@@ -50,16 +56,21 @@ export const useNonFungibleIds = (resourceId: string, addresses: string[]) => {
 			const aggregatedResult = responses.reduce(
 				(accumulator, response) => {
 					accumulator.items.push(...response.items)
-					accumulator.next_cursors.push(response.next_cursor || null)
+					accumulator.pageParams.push({
+						next_cursor: response.next_cursor,
+						ledger_state: response.ledger_state,
+					})
 					return accumulator
 				},
-				{ items: [], next_cursors: [] },
+				{ items: [], pageParams: [] },
 			)
 
 			return aggregatedResult
 		},
 		getNextPageParam: lastPage =>
-			lastPage.next_cursors?.filter(cursor => cursor !== null).length > 0 ? lastPage.next_cursors : undefined,
+			lastPage.pageParams?.filter(({ next_cursor }) => next_cursor !== null).length > 0
+				? lastPage.pageParams
+				: undefined,
 		enabled: !!resourceId && !!state && !isLoading && vaults.length > 0,
 	})
 }
