+++ packages/ui/src/pages/transfer/swap/styles.css.ts
@@ -1,7 +1,6 @@
 import { globalStyle, style } from '@vanilla-extract/css'
 
 import { sprinkles } from 'ui/src/components/system/sprinkles.css'
-import { vars } from 'ui/src/components/system/theme.css'
 
 export const swapDropdownContentWrapper = style([
 	{
